further split up config_classes
