+++ slk/config/network.py
@@ -1,41 +1,13 @@
 from __future__ import annotations
 
-from dataclasses import dataclass
-from typing import List, Optional, Union
+from typing import List, Optional
 
 from xrpl import CryptoAlgorithm
 from xrpl.core.addresscodec import encode_account_public_key, encode_node_public_key
 from xrpl.core.keypairs import derive_keypair, generate_seed
-from xrpl.models import Amount
 from xrpl.wallet import Wallet
 
-from slk.common import same_amount_new_value
-
-
-@dataclass
-class Keypair:
-    public_key: str
-    secret_key: str
-    account_id: Optional[str]
-
-
-class Ports:
-    """
-    Port numbers for various services.
-    Port numbers differ by cfg_index so different configs can run
-    at the same time without interfering with each other.
-    """
-
-    peer_port_base = 51235
-    http_admin_port_base = 5005
-    ws_public_port_base = 6005
-
-    def __init__(self, cfg_index: int):
-        self.peer_port = Ports.peer_port_base + cfg_index
-        self.http_admin_port = Ports.http_admin_port_base + cfg_index
-        self.ws_public_port = Ports.ws_public_port_base + (2 * cfg_index)
-        # note admin port uses public port base
-        self.ws_admin_port = Ports.ws_public_port_base + (2 * cfg_index) + 1
+from slk.config.config_classes import Keypair, Ports
 
 
 class Network:
@@ -87,23 +59,3 @@ class SidechainNetwork(Network):
                 )
             )
         return result
-
-
-class XChainAsset:
-    def __init__(
-        self,
-        main_asset: Amount,
-        side_asset: Amount,
-        main_value: Union[int, float],
-        side_value: Union[int, float],
-        main_refund_penalty: Union[int, float],
-        side_refund_penalty: Union[int, float],
-    ):
-        self.main_asset = same_amount_new_value(main_asset, main_value)
-        self.side_asset = same_amount_new_value(side_asset, side_value)
-        self.main_refund_penalty = same_amount_new_value(
-            main_asset, main_refund_penalty
-        )
-        self.side_refund_penalty = same_amount_new_value(
-            side_asset, side_refund_penalty
-        )
